(Added framework classes. Added player sprites. Added player entity.)
Basketball is an interesting arcade game to clone because it is one of the first to use sprites. Because of this, the cloned game will also move away from drawing lines and rectangles directly to the framebuffer, instead opting for drawing actual sprites. To this effect a `Sprite` class was created which extends `Entity` with a `graphic` property. This property stores an image of the sprite which is automatically drawn by the sprite's `draw()` call. The sprite's width and height properties are determined by the graphic.

Basketball features two teams of two players with four players total. These teams are represented by groups of `Player` classes. All of the players in a team move simultaneously so no special movement or input logic is necessary.

Like with font assets in the previous projects (kha-pong and kha-astro-race), the player sprites have to be loaded and prepared by `kha.Assets` before creating a new instance of the game. This is of course done in the `Main` class.