(Added court and limited player movement. Renamed assets. Removed unused imports.)
Players can only move within a vertical boundary. This vertical boundary is set per player instance via the `setMovementBounds()` method.

The court or playing field is a rectangular area. It could have been represented by a simple `fillRect()` from Kha's graphics API, but to keep things sensible it is given its own class. The court is made an `Object` instead of an `Entity` because it is not exactly necessary to perform collision checks against it. It is much simpler to set movement boundaries for the ball and player instead.

Also, the court is drawn in a different colour because of potential conflicts with the ball and player sprites. And it's prettier that way.