(Added ball. Updated visuals to better match original arcade game.)
Basketball is, truthfully, just a four-paddle variation on Pong with a limited score area. The ball behaves generally the same in both games save for a few special collision cases that are going to be handled once baskets are implemented. Ball movement is bounded inside a rectangular area, and the ball "bounces" off of the area's invisible walls.

The visuals are also updated to adhere to the style of the original arcade game.