(Added player-ball collision.)
Basketball's player-ball collision is slightly different than paddle-ball collision in Pong. The players on the left team throw the ball to the right side of the court, and the players on the right team throw the ball to the left side. Also, if the ball collides with the lower half of the player, it bounces towards the court instead of the cieling.

Because the players often move at the same time as the ball collides on top or below them, there needs to exist a mechanism that only resolves collision the first time the ball overlaps with the player. Otherwise the collision results may be unpredictable.

The `Entity` class is given a `isOverlapping` flag to facilitate that mechanism (which is a simple `resolvedCollision` flag itself). The `isOverlapping` flag is false by default and remains false unless the `overlapsEntity()` or `overlapsPoint()` methods return true. This way checks can be made to figure out when collision (overlaps) first happened and when they stopped happening.