(Updated basket graphic. Added ball-basket collision. Refactored some code.)
Resolving ball-basket collision is not as difficult as it seems. If the ball is moving downwards when the collision occurs, check if it is inside the basket. If it is, set vertical direction of movement to zero. If it isn't, bounce the ball off the basket. On the other hand, the ball can collide with the basket while it is moving upwards. In that case the ball simply bounces off the basket by changing its horizontal direction of movement.