(Added font asset. Added UI elements: timer and score displays. Added end-game behaviour.)
Displaying text requires some boilerplate, which is why the `gmae.ui.Text` class was created. A `Text` is an `Object` that has a string value, font, and font size. The text's width and height properties are automatically calculated from the text value. Score displays are instances of the `Text` class, but the timer is placed in its own class because it is a little different.

The timer has, well, a set time that it counts down from and an `onComplete` callback. It also formats the text output of the current time. Once the time runs out and reaches zero, the callback function is executed. A `TimeTask` from `kha.Scheduler` updates the timer every second for the timer's specified duration.

The `onComplete` callback is set to the `endGame()` function defined inside the `Project` class which 'deactivates' the ball by setting the `active` and `visible` flags to false. These flags were added to `Object` to make such behaviour possible.